Struct aws_sdk_connect::client::Client

pub struct Client<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

An ergonomic service client for AmazonConnectService.

This client allows ergonomic access to a AmazonConnectService-shaped service. Each method corresponds to an endpoint defined in the service’s Smithy model, and the request and response shapes are auto-generated from that same model.

Using a Client

Once you have a client set up, you can access the service’s endpoints by calling the appropriate method on Client. Each such method returns a request builder for that endpoint, with methods for setting the various fields of the request. Once your request is complete, use the send method to send the request. send returns a future, which you then have to .await to get the service’s response.

Implementations

impl<C, M, R> Client<C, M, R>

pub fn with_config(client: Client<C, M, R>, conf: Config) -> Self

pub fn conf(&self) -> &Config

impl<C, M, R> Client<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub fn associate_approved_origin(&self) -> AssociateApprovedOrigin<C, M, R>

pub fn associate_bot(&self) -> AssociateBot<C, M, R>

pub fn associate_instance_storage_config(
&self
) -> AssociateInstanceStorageConfig<C, M, R>

pub fn associate_lambda_function(&self) -> AssociateLambdaFunction<C, M, R>

pub fn associate_lex_bot(&self) -> AssociateLexBot<C, M, R>

pub fn associate_queue_quick_connects(
&self
) -> AssociateQueueQuickConnects<C, M, R>

pub fn associate_routing_profile_queues(
&self
) -> AssociateRoutingProfileQueues<C, M, R>

pub fn associate_security_key(&self) -> AssociateSecurityKey<C, M, R>

pub fn create_agent_status(&self) -> CreateAgentStatus<C, M, R>

pub fn create_contact_flow(&self) -> CreateContactFlow<C, M, R>

pub fn create_hours_of_operation(&self) -> CreateHoursOfOperation<C, M, R>

pub fn create_instance(&self) -> CreateInstance<C, M, R>

pub fn create_integration_association(
&self
) -> CreateIntegrationAssociation<C, M, R>

pub fn create_queue(&self) -> CreateQueue<C, M, R>

pub fn create_quick_connect(&self) -> CreateQuickConnect<C, M, R>

pub fn create_routing_profile(&self) -> CreateRoutingProfile<C, M, R>

pub fn create_use_case(&self) -> CreateUseCase<C, M, R>

pub fn create_user(&self) -> CreateUser<C, M, R>

pub fn create_user_hierarchy_group(&self) -> CreateUserHierarchyGroup<C, M, R>

pub fn delete_hours_of_operation(&self) -> DeleteHoursOfOperation<C, M, R>

pub fn delete_instance(&self) -> DeleteInstance<C, M, R>

pub fn delete_integration_association(
&self
) -> DeleteIntegrationAssociation<C, M, R>

pub fn delete_quick_connect(&self) -> DeleteQuickConnect<C, M, R>

pub fn delete_use_case(&self) -> DeleteUseCase<C, M, R>

pub fn delete_user(&self) -> DeleteUser<C, M, R>

pub fn delete_user_hierarchy_group(&self) -> DeleteUserHierarchyGroup<C, M, R>

pub fn describe_agent_status(&self) -> DescribeAgentStatus<C, M, R>

pub fn describe_contact_flow(&self) -> DescribeContactFlow<C, M, R>

pub fn describe_hours_of_operation(&self) -> DescribeHoursOfOperation<C, M, R>

pub fn describe_instance(&self) -> DescribeInstance<C, M, R>

pub fn describe_instance_attribute(&self) -> DescribeInstanceAttribute<C, M, R>

pub fn describe_instance_storage_config(
&self
) -> DescribeInstanceStorageConfig<C, M, R>

pub fn describe_queue(&self) -> DescribeQueue<C, M, R>

pub fn describe_quick_connect(&self) -> DescribeQuickConnect<C, M, R>

pub fn describe_routing_profile(&self) -> DescribeRoutingProfile<C, M, R>

pub fn describe_user(&self) -> DescribeUser<C, M, R>

pub fn describe_user_hierarchy_group(
&self
) -> DescribeUserHierarchyGroup<C, M, R>

pub fn describe_user_hierarchy_structure(
&self
) -> DescribeUserHierarchyStructure<C, M, R>

pub fn disassociate_approved_origin(
&self
) -> DisassociateApprovedOrigin<C, M, R>

pub fn disassociate_bot(&self) -> DisassociateBot<C, M, R>

pub fn disassociate_instance_storage_config(
&self
) -> DisassociateInstanceStorageConfig<C, M, R>

pub fn disassociate_lambda_function(
&self
) -> DisassociateLambdaFunction<C, M, R>

pub fn disassociate_lex_bot(&self) -> DisassociateLexBot<C, M, R>

pub fn disassociate_queue_quick_connects(
&self
) -> DisassociateQueueQuickConnects<C, M, R>

pub fn disassociate_routing_profile_queues(
&self
) -> DisassociateRoutingProfileQueues<C, M, R>

pub fn disassociate_security_key(&self) -> DisassociateSecurityKey<C, M, R>

pub fn get_contact_attributes(&self) -> GetContactAttributes<C, M, R>

pub fn get_current_metric_data(&self) -> GetCurrentMetricData<C, M, R>

pub fn get_federation_token(&self) -> GetFederationToken<C, M, R>

pub fn get_metric_data(&self) -> GetMetricData<C, M, R>

pub fn list_agent_statuses(&self) -> ListAgentStatuses<C, M, R>

pub fn list_approved_origins(&self) -> ListApprovedOrigins<C, M, R>

pub fn list_bots(&self) -> ListBots<C, M, R>

pub fn list_contact_flows(&self) -> ListContactFlows<C, M, R>

pub fn list_hours_of_operations(&self) -> ListHoursOfOperations<C, M, R>

pub fn list_instance_attributes(&self) -> ListInstanceAttributes<C, M, R>

pub fn list_instances(&self) -> ListInstances<C, M, R>

pub fn list_instance_storage_configs(
&self
) -> ListInstanceStorageConfigs<C, M, R>

pub fn list_integration_associations(
&self
) -> ListIntegrationAssociations<C, M, R>

pub fn list_lambda_functions(&self) -> ListLambdaFunctions<C, M, R>

pub fn list_lex_bots(&self) -> ListLexBots<C, M, R>

pub fn list_phone_numbers(&self) -> ListPhoneNumbers<C, M, R>

pub fn list_prompts(&self) -> ListPrompts<C, M, R>

pub fn list_queue_quick_connects(&self) -> ListQueueQuickConnects<C, M, R>

pub fn list_queues(&self) -> ListQueues<C, M, R>

pub fn list_quick_connects(&self) -> ListQuickConnects<C, M, R>

pub fn list_routing_profile_queues(&self) -> ListRoutingProfileQueues<C, M, R>

pub fn list_routing_profiles(&self) -> ListRoutingProfiles<C, M, R>

pub fn list_security_keys(&self) -> ListSecurityKeys<C, M, R>

pub fn list_security_profiles(&self) -> ListSecurityProfiles<C, M, R>

pub fn list_tags_for_resource(&self) -> ListTagsForResource<C, M, R>

pub fn list_use_cases(&self) -> ListUseCases<C, M, R>

pub fn list_user_hierarchy_groups(&self) -> ListUserHierarchyGroups<C, M, R>

pub fn list_users(&self) -> ListUsers<C, M, R>

pub fn resume_contact_recording(&self) -> ResumeContactRecording<C, M, R>

pub fn start_chat_contact(&self) -> StartChatContact<C, M, R>

pub fn start_contact_recording(&self) -> StartContactRecording<C, M, R>

pub fn start_outbound_voice_contact(&self) -> StartOutboundVoiceContact<C, M, R>

pub fn start_task_contact(&self) -> StartTaskContact<C, M, R>

pub fn stop_contact(&self) -> StopContact<C, M, R>

pub fn stop_contact_recording(&self) -> StopContactRecording<C, M, R>

pub fn suspend_contact_recording(&self) -> SuspendContactRecording<C, M, R>

pub fn tag_resource(&self) -> TagResource<C, M, R>

pub fn untag_resource(&self) -> UntagResource<C, M, R>

pub fn update_agent_status(&self) -> UpdateAgentStatus<C, M, R>

pub fn update_contact_attributes(&self) -> UpdateContactAttributes<C, M, R>

pub fn update_contact_flow_content(&self) -> UpdateContactFlowContent<C, M, R>

pub fn update_contact_flow_name(&self) -> UpdateContactFlowName<C, M, R>

pub fn update_hours_of_operation(&self) -> UpdateHoursOfOperation<C, M, R>

pub fn update_instance_attribute(&self) -> UpdateInstanceAttribute<C, M, R>

pub fn update_instance_storage_config(
&self
) -> UpdateInstanceStorageConfig<C, M, R>

pub fn update_queue_hours_of_operation(
&self
) -> UpdateQueueHoursOfOperation<C, M, R>

pub fn update_queue_max_contacts(&self) -> UpdateQueueMaxContacts<C, M, R>

pub fn update_queue_name(&self) -> UpdateQueueName<C, M, R>

pub fn update_queue_outbound_caller_config(
&self
) -> UpdateQueueOutboundCallerConfig<C, M, R>

pub fn update_queue_status(&self) -> UpdateQueueStatus<C, M, R>

pub fn update_quick_connect_config(&self) -> UpdateQuickConnectConfig<C, M, R>

pub fn update_quick_connect_name(&self) -> UpdateQuickConnectName<C, M, R>

pub fn update_routing_profile_concurrency(
&self
) -> UpdateRoutingProfileConcurrency<C, M, R>

pub fn update_routing_profile_default_outbound_queue(
&self
) -> UpdateRoutingProfileDefaultOutboundQueue<C, M, R>

pub fn update_routing_profile_name(&self) -> UpdateRoutingProfileName<C, M, R>

pub fn update_routing_profile_queues(
&self
) -> UpdateRoutingProfileQueues<C, M, R>

pub fn update_user_hierarchy(&self) -> UpdateUserHierarchy<C, M, R>

pub fn update_user_hierarchy_group_name(
&self
) -> UpdateUserHierarchyGroupName<C, M, R>

pub fn update_user_hierarchy_structure(
&self
) -> UpdateUserHierarchyStructure<C, M, R>

pub fn update_user_identity_info(&self) -> UpdateUserIdentityInfo<C, M, R>

pub fn update_user_phone_config(&self) -> UpdateUserPhoneConfig<C, M, R>

pub fn update_user_routing_profile(&self) -> UpdateUserRoutingProfile<C, M, R>

pub fn update_user_security_profiles(
&self
) -> UpdateUserSecurityProfiles<C, M, R>

impl<C> Client<C, AwsMiddleware, Standard>

pub fn from_conf_conn(conf: Config, conn: C) -> Self

impl Client<DynConnector, AwsMiddleware, Standard>

pub fn new(config: &Config) -> Self

pub fn from_conf(conf: Config) -> Self