Module aws_sdk_connect::error

Modules

contact_flow_not_published_exception

See ContactFlowNotPublishedException

contact_not_found_exception

See ContactNotFoundException

destination_not_allowed_exception

See DestinationNotAllowedException

duplicate_resource_exception

See DuplicateResourceException

internal_service_exception

See InternalServiceException

invalid_contact_flow_exception

See InvalidContactFlowException

invalid_parameter_exception

See InvalidParameterException

invalid_request_exception

See InvalidRequestException

limit_exceeded_exception

See LimitExceededException

outbound_contact_not_permitted_exception

See OutboundContactNotPermittedException

resource_conflict_exception

See ResourceConflictException

resource_in_use_exception

See ResourceInUseException

resource_not_found_exception

See ResourceNotFoundException

service_quota_exceeded_exception

See ServiceQuotaExceededException

throttling_exception

See ThrottlingException

user_not_found_exception

See UserNotFoundException

Structs

AssociateApprovedOriginError

AssociateBotError

AssociateInstanceStorageConfigError

AssociateLambdaFunctionError

AssociateLexBotError

AssociateQueueQuickConnectsError

AssociateRoutingProfileQueuesError

AssociateSecurityKeyError

ContactFlowNotPublishedException

The contact flow has not been published.

ContactNotFoundException

The contact with the specified ID is not active or does not exist.

CreateAgentStatusError

CreateContactFlowError

CreateHoursOfOperationError

CreateInstanceError

CreateIntegrationAssociationError

CreateQueueError

CreateQuickConnectError

CreateRoutingProfileError

CreateUseCaseError

CreateUserError

CreateUserHierarchyGroupError

DeleteHoursOfOperationError

DeleteInstanceError

DeleteIntegrationAssociationError

DeleteQuickConnectError

DeleteUseCaseError

DeleteUserError

DeleteUserHierarchyGroupError

DescribeAgentStatusError

DescribeContactFlowError

DescribeHoursOfOperationError

DescribeInstanceAttributeError

DescribeInstanceError

DescribeInstanceStorageConfigError

DescribeQueueError

DescribeQuickConnectError

DescribeRoutingProfileError

DescribeUserError

DescribeUserHierarchyGroupError

DescribeUserHierarchyStructureError

DestinationNotAllowedException

Outbound calls to the destination number are not allowed.

DisassociateApprovedOriginError

DisassociateBotError

DisassociateInstanceStorageConfigError

DisassociateLambdaFunctionError

DisassociateLexBotError

DisassociateQueueQuickConnectsError

DisassociateRoutingProfileQueuesError

DisassociateSecurityKeyError

DuplicateResourceException

A resource with the specified name already exists.

GetContactAttributesError

GetCurrentMetricDataError

GetFederationTokenError

GetMetricDataError

InternalServiceException

Request processing failed because of an error or failure with the service.

InvalidContactFlowException

The contact flow is not valid.

InvalidParameterException

One or more of the specified parameters are not valid.

InvalidRequestException

The request is not valid.

LimitExceededException

The allowed limit for the resource has been exceeded.

ListAgentStatusesError

ListApprovedOriginsError

ListBotsError

ListContactFlowsError

ListHoursOfOperationsError

ListInstanceAttributesError

ListInstanceStorageConfigsError

ListInstancesError

ListIntegrationAssociationsError

ListLambdaFunctionsError

ListLexBotsError

ListPhoneNumbersError

ListPromptsError

ListQueueQuickConnectsError

ListQueuesError

ListQuickConnectsError

ListRoutingProfileQueuesError

ListRoutingProfilesError

ListSecurityKeysError

ListSecurityProfilesError

ListTagsForResourceError

ListUseCasesError

ListUserHierarchyGroupsError

ListUsersError

OutboundContactNotPermittedException

The contact is not permitted.

ResourceConflictException

A resource already has that name.

ResourceInUseException

That resource is already in use. Please try another.

ResourceNotFoundException

The specified resource was not found.

ResumeContactRecordingError

ServiceQuotaExceededException

The service quota has been exceeded.

StartChatContactError

StartContactRecordingError

StartOutboundVoiceContactError

StartTaskContactError

StopContactError

StopContactRecordingError

SuspendContactRecordingError

TagResourceError

ThrottlingException

The throttling limit has been exceeded.

UntagResourceError

UpdateAgentStatusError

UpdateContactAttributesError

UpdateContactFlowContentError

UpdateContactFlowNameError

UpdateHoursOfOperationError

UpdateInstanceAttributeError

UpdateInstanceStorageConfigError

UpdateQueueHoursOfOperationError

UpdateQueueMaxContactsError

UpdateQueueNameError

UpdateQueueOutboundCallerConfigError

UpdateQueueStatusError

UpdateQuickConnectConfigError

UpdateQuickConnectNameError

UpdateRoutingProfileConcurrencyError

UpdateRoutingProfileDefaultOutboundQueueError

UpdateRoutingProfileNameError

UpdateRoutingProfileQueuesError

UpdateUserHierarchyError

UpdateUserHierarchyGroupNameError

UpdateUserHierarchyStructureError

UpdateUserIdentityInfoError

UpdateUserPhoneConfigError

UpdateUserRoutingProfileError

UpdateUserSecurityProfilesError

UserNotFoundException

No user with the specified credentials was found in the Amazon Connect instance.

Enums

AssociateApprovedOriginErrorKind

AssociateBotErrorKind

AssociateInstanceStorageConfigErrorKind

AssociateLambdaFunctionErrorKind

AssociateLexBotErrorKind

AssociateQueueQuickConnectsErrorKind

AssociateRoutingProfileQueuesErrorKind

AssociateSecurityKeyErrorKind

CreateAgentStatusErrorKind

CreateContactFlowErrorKind

CreateHoursOfOperationErrorKind

CreateInstanceErrorKind

CreateIntegrationAssociationErrorKind

CreateQueueErrorKind

CreateQuickConnectErrorKind

CreateRoutingProfileErrorKind

CreateUseCaseErrorKind

CreateUserErrorKind

CreateUserHierarchyGroupErrorKind

DeleteHoursOfOperationErrorKind

DeleteInstanceErrorKind

DeleteIntegrationAssociationErrorKind

DeleteQuickConnectErrorKind

DeleteUseCaseErrorKind

DeleteUserErrorKind

DeleteUserHierarchyGroupErrorKind

DescribeAgentStatusErrorKind

DescribeContactFlowErrorKind

DescribeHoursOfOperationErrorKind

DescribeInstanceAttributeErrorKind

DescribeInstanceErrorKind

DescribeInstanceStorageConfigErrorKind

DescribeQueueErrorKind

DescribeQuickConnectErrorKind

DescribeRoutingProfileErrorKind

DescribeUserErrorKind

DescribeUserHierarchyGroupErrorKind

DescribeUserHierarchyStructureErrorKind

DisassociateApprovedOriginErrorKind

DisassociateBotErrorKind

DisassociateInstanceStorageConfigErrorKind

DisassociateLambdaFunctionErrorKind

DisassociateLexBotErrorKind

DisassociateQueueQuickConnectsErrorKind

DisassociateRoutingProfileQueuesErrorKind

DisassociateSecurityKeyErrorKind

GetContactAttributesErrorKind

GetCurrentMetricDataErrorKind

GetFederationTokenErrorKind

GetMetricDataErrorKind

ListAgentStatusesErrorKind

ListApprovedOriginsErrorKind

ListBotsErrorKind

ListContactFlowsErrorKind

ListHoursOfOperationsErrorKind

ListInstanceAttributesErrorKind

ListInstanceStorageConfigsErrorKind

ListInstancesErrorKind

ListIntegrationAssociationsErrorKind

ListLambdaFunctionsErrorKind

ListLexBotsErrorKind

ListPhoneNumbersErrorKind

ListPromptsErrorKind

ListQueueQuickConnectsErrorKind

ListQueuesErrorKind

ListQuickConnectsErrorKind

ListRoutingProfileQueuesErrorKind

ListRoutingProfilesErrorKind

ListSecurityKeysErrorKind

ListSecurityProfilesErrorKind

ListTagsForResourceErrorKind

ListUseCasesErrorKind

ListUserHierarchyGroupsErrorKind

ListUsersErrorKind

ResumeContactRecordingErrorKind

StartChatContactErrorKind

StartContactRecordingErrorKind

StartOutboundVoiceContactErrorKind

StartTaskContactErrorKind

StopContactErrorKind

StopContactRecordingErrorKind

SuspendContactRecordingErrorKind

TagResourceErrorKind

UntagResourceErrorKind

UpdateAgentStatusErrorKind

UpdateContactAttributesErrorKind

UpdateContactFlowContentErrorKind

UpdateContactFlowNameErrorKind

UpdateHoursOfOperationErrorKind

UpdateInstanceAttributeErrorKind

UpdateInstanceStorageConfigErrorKind

UpdateQueueHoursOfOperationErrorKind

UpdateQueueMaxContactsErrorKind

UpdateQueueNameErrorKind

UpdateQueueOutboundCallerConfigErrorKind

UpdateQueueStatusErrorKind

UpdateQuickConnectConfigErrorKind

UpdateQuickConnectNameErrorKind

UpdateRoutingProfileConcurrencyErrorKind

UpdateRoutingProfileDefaultOutboundQueueErrorKind

UpdateRoutingProfileNameErrorKind

UpdateRoutingProfileQueuesErrorKind

UpdateUserHierarchyErrorKind

UpdateUserHierarchyGroupNameErrorKind

UpdateUserHierarchyStructureErrorKind

UpdateUserIdentityInfoErrorKind

UpdateUserPhoneConfigErrorKind

UpdateUserRoutingProfileErrorKind

UpdateUserSecurityProfilesErrorKind