Module aws_sdk_connect::model

Modules

agent_status

See AgentStatus

agent_status_summary

See AgentStatusSummary

attribute

See Attribute

chat_message

See ChatMessage

contact_flow

See ContactFlow

contact_flow_summary

See ContactFlowSummary

credentials

See Credentials

current_metric

See CurrentMetric

current_metric_data

See CurrentMetricData

current_metric_result

See CurrentMetricResult

dimensions

See Dimensions

encryption_config

See EncryptionConfig

filters

See Filters

hierarchy_group

See HierarchyGroup

hierarchy_group_summary

See HierarchyGroupSummary

hierarchy_level

See HierarchyLevel

hierarchy_level_update

See HierarchyLevelUpdate

hierarchy_path

See HierarchyPath

hierarchy_structure

See HierarchyStructure

hierarchy_structure_update

See HierarchyStructureUpdate

historical_metric

See HistoricalMetric

historical_metric_data

See HistoricalMetricData

historical_metric_result

See HistoricalMetricResult

hours_of_operation

See HoursOfOperation

hours_of_operation_config

See HoursOfOperationConfig

hours_of_operation_summary

See HoursOfOperationSummary

hours_of_operation_time_slice

See HoursOfOperationTimeSlice

instance

See Instance

instance_status_reason

See InstanceStatusReason

instance_storage_config

See InstanceStorageConfig

instance_summary

See InstanceSummary

integration_association_summary

See IntegrationAssociationSummary

kinesis_firehose_config

See KinesisFirehoseConfig

kinesis_stream_config

See KinesisStreamConfig

kinesis_video_stream_config

See KinesisVideoStreamConfig

lex_bot

See LexBot

lex_bot_config

See LexBotConfig

lex_v2_bot

See LexV2Bot

media_concurrency

See MediaConcurrency

outbound_caller_config

See OutboundCallerConfig

participant_details

See ParticipantDetails

phone_number_quick_connect_config

See PhoneNumberQuickConnectConfig

phone_number_summary

See PhoneNumberSummary

problem_detail

See ProblemDetail

prompt_summary

See PromptSummary

queue

See Queue

queue_quick_connect_config

See QueueQuickConnectConfig

queue_reference

See QueueReference

queue_summary

See QueueSummary

quick_connect

See QuickConnect

quick_connect_config

See QuickConnectConfig

quick_connect_summary

See QuickConnectSummary

reference

See Reference

routing_profile

See RoutingProfile

routing_profile_queue_config

See RoutingProfileQueueConfig

routing_profile_queue_config_summary

See RoutingProfileQueueConfigSummary

routing_profile_queue_reference

See RoutingProfileQueueReference

routing_profile_summary

See RoutingProfileSummary

s3_config

See S3Config

security_key

See SecurityKey

security_profile_summary

See SecurityProfileSummary

threshold

See Threshold

use_case

See UseCase

user

See User

user_identity_info

See UserIdentityInfo

user_phone_config

See UserPhoneConfig

user_quick_connect_config

See UserQuickConnectConfig

user_summary

See UserSummary

voice_recording_configuration

See VoiceRecordingConfiguration

Structs

AgentStatus

Contains information about an agent status.

AgentStatusSummary

Summary information for an agent status.

Attribute

A toggle for an individual feature at the instance level.

ChatMessage

A chat message.

ContactFlow

Contains information about a contact flow.

ContactFlowSummary

Contains summary information about a contact flow.

Credentials

Contains credentials to use for federation.

CurrentMetric

Contains information about a real-time metric. For a description of each metric, see Real-time Metrics Definitions in the Amazon Connect Administrator

CurrentMetricData

Contains the data for a real-time metric.

CurrentMetricResult

Contains information about a set of real-time metrics.

Dimensions

Contains information about the dimensions for a set of metrics.

EncryptionConfig

The encryption configuration.

Filters

Contains the filter to apply when retrieving metrics.

HierarchyGroup

Contains information about a hierarchy group.

HierarchyGroupSummary

Contains summary information about a hierarchy group.

HierarchyLevel

Contains information about a hierarchy level.

HierarchyLevelUpdate

Contains information about the hierarchy level to update.

HierarchyPath

Contains information about the levels of a hierarchy group.

HierarchyStructure

Contains information about a hierarchy structure.

HierarchyStructureUpdate

Contains information about the level hierarchy to update.

HistoricalMetric

Contains information about a historical metric. For a description of each metric, see Historical Metrics Definitions in the Amazon Connect Administrator

HistoricalMetricData

Contains the data for a historical metric.

HistoricalMetricResult

Contains information about the historical metrics retrieved.

HoursOfOperation

Information about of the hours of operation.

HoursOfOperationConfig

Contains information about the hours of operation.

HoursOfOperationSummary

Contains summary information about hours of operation for a contact center.

HoursOfOperationTimeSlice

The start time or end time for an hours of operation.

Instance

The Amazon Connect instance.

InstanceStatusReason

Relevant

InstanceStorageConfig

The storage configuration for the instance.

InstanceSummary

Information about the instance.

IntegrationAssociationSummary

Contains summary information about the associated AppIntegrations.

KinesisFirehoseConfig

Configuration information of a Kinesis Data Firehose delivery stream.

KinesisStreamConfig

Configuration information of a Kinesis data stream.

KinesisVideoStreamConfig

Configuration information of a Kinesis video stream.

LexBot

Configuration information of an Amazon Lex bot.

LexBotConfig

Configuration information of an Amazon Lex or Amazon Lex V2 bot.

LexV2Bot

Configuration information of an Amazon Lex V2 bot.

MediaConcurrency

Contains information about which channels are supported, and how many contacts an agent can

OutboundCallerConfig

The outbound caller ID name, number, and outbound whisper flow.

ParticipantDetails

The customer's details.

PhoneNumberQuickConnectConfig

Contains information about a phone number for a quick connect.

PhoneNumberSummary

Contains summary information about a phone number for a contact center.

ProblemDetail

Information about a problem detail.

PromptSummary

Contains information about the prompt.

Queue

Contains information about a queue.

QueueQuickConnectConfig

Contains information about a queue for a quick connect. The contact flow must be of type

QueueReference

Contains information about a queue resource for which metrics are returned.

QueueSummary

Contains summary information about a queue.

QuickConnect

Contains information about a quick connect.

QuickConnectConfig

Contains configuration settings for a quick connect.

QuickConnectSummary

Contains summary information about a quick connect.

Reference

A link that an agent selects to complete a given task. You can have up to 4,096 UTF-8 bytes

RoutingProfile

Contains information about a routing profile.

RoutingProfileQueueConfig

Contains information about the queue and channel for which priority and delay can be

RoutingProfileQueueConfigSummary

Contains summary information about a routing profile queue.

RoutingProfileQueueReference

Contains the channel and queue identifier for a routing profile.

RoutingProfileSummary

Contains summary information about a routing profile.

S3Config

Information about the Amazon Simple Storage Service (Amazon S3) storage type.

SecurityKey

Configuration information of the security key.

SecurityProfileSummary

Contains information about a security profile.

Threshold

Contains information about the threshold for service level metrics.

UseCase

Contains the

User

Contains information about a user account for a Amazon Connect instance.

UserIdentityInfo

Contains information about the identity of a user.

UserPhoneConfig

Contains information about the phone configuration settings for a user.

UserQuickConnectConfig

Contains information about the quick connect configuration settings for a user. The contact

UserSummary

Contains summary information about a user.

VoiceRecordingConfiguration

Contains information about the recording configuration settings.

Enums

AgentStatusState

AgentStatusType

Channel

Comparison

ContactFlowType

CurrentMetricName

The current metric names.

DirectoryType

EncryptionType

Grouping

HistoricalMetricName

The historical metric names.

HoursOfOperationDays

InstanceAttributeType

InstanceStatus

InstanceStorageResourceType

IntegrationType

LexVersion

PhoneNumberCountryCode

PhoneNumberType

PhoneType

QueueStatus

QueueType

QuickConnectType

ReferenceType

ResourceType

SourceType

Statistic

StorageType

Unit

UseCaseType

VoiceRecordingTrack