Module aws_sdk_connect::operation

Structs

AssociateApprovedOrigin

This API is in preview release for Amazon Connect and is subject to change.

AssociateBot

This API is in preview release for Amazon Connect and is subject to change.

AssociateInstanceStorageConfig

This API is in preview release for Amazon Connect and is subject to change.

AssociateLambdaFunction

This API is in preview release for Amazon Connect and is subject to change.

AssociateLexBot

This API is in preview release for Amazon Connect and is subject to change.

AssociateQueueQuickConnects

This API is in preview release for Amazon Connect and is subject to change.

AssociateRoutingProfileQueues

Associates a set of queues with a routing profile.

AssociateSecurityKey

This API is in preview release for Amazon Connect and is subject to change.

CreateAgentStatus

This API is in preview release for Amazon Connect and is subject to change.

CreateContactFlow

Creates a contact flow for the specified Amazon Connect instance.

CreateHoursOfOperation

This API is in preview release for Amazon Connect and is subject to change.

CreateInstance

This API is in preview release for Amazon Connect and is subject to change.

CreateIntegrationAssociation

Create an AppIntegration association with an Amazon Connect instance.

CreateQueue

This API is in preview release for Amazon Connect and is subject to change.

CreateQuickConnect

Creates a quick connect for the specified Amazon Connect instance.

CreateRoutingProfile

Creates a new routing profile.

CreateUseCase

Creates a use case for an AppIntegration association.

CreateUser

Creates a user account for the specified Amazon Connect instance.

CreateUserHierarchyGroup

Creates a new user hierarchy group.

DeleteHoursOfOperation

This API is in preview release for Amazon Connect and is subject to change.

DeleteInstance

This API is in preview release for Amazon Connect and is subject to change.

DeleteIntegrationAssociation

Deletes an AppIntegration association from an Amazon Connect instance. The association must not have

DeleteQuickConnect

Deletes a quick connect.

DeleteUseCase

Deletes a use case from an AppIntegration association.

DeleteUser

Deletes a user account from the specified Amazon Connect instance.

DeleteUserHierarchyGroup

Deletes an existing user hierarchy group. It must not be associated with any agents or have

DescribeAgentStatus

This API is in preview release for Amazon Connect and is subject to change.

DescribeContactFlow

Describes the specified contact flow.

DescribeHoursOfOperation

This API is in preview release for Amazon Connect and is subject to change.

DescribeInstance

This API is in preview release for Amazon Connect and is subject to change.

DescribeInstanceAttribute

This API is in preview release for Amazon Connect and is subject to change.

DescribeInstanceStorageConfig

This API is in preview release for Amazon Connect and is subject to change.

DescribeQueue

This API is in preview release for Amazon Connect and is subject to change.

DescribeQuickConnect

Describes the quick connect.

DescribeRoutingProfile

Describes the specified routing profile.

DescribeUser

Describes the specified user account. You can find the instance ID in the console (it’s the

DescribeUserHierarchyGroup

Describes the specified hierarchy group.

DescribeUserHierarchyStructure

Describes the hierarchy structure of the specified Amazon Connect instance.

DisassociateApprovedOrigin

This API is in preview release for Amazon Connect and is subject to change.

DisassociateBot

This API is in preview release for Amazon Connect and is subject to change.

DisassociateInstanceStorageConfig

This API is in preview release for Amazon Connect and is subject to change.

DisassociateLambdaFunction

This API is in preview release for Amazon Connect and is subject to change.

DisassociateLexBot

This API is in preview release for Amazon Connect and is subject to change.

DisassociateQueueQuickConnects

This API is in preview release for Amazon Connect and is subject to change.

DisassociateRoutingProfileQueues

Disassociates a set of queues from a routing profile.

DisassociateSecurityKey

This API is in preview release for Amazon Connect and is subject to change.

GetContactAttributes

Retrieves the contact attributes for the specified contact.

GetCurrentMetricData

Gets the real-time metric data from the specified Amazon Connect instance.

GetFederationToken

Retrieves a token for federation.

GetMetricData

Gets historical metric data from the specified Amazon Connect instance.

ListAgentStatuses

This API is in preview release for Amazon Connect and is subject to change.

ListApprovedOrigins

This API is in preview release for Amazon Connect and is subject to change.

ListBots

This API is in preview release for Amazon Connect and is subject to change.

ListContactFlows

Provides information about the contact flows for the specified Amazon Connect instance.

ListHoursOfOperations

Provides information about the hours of operation for the specified Amazon Connect instance.

ListInstanceAttributes

This API is in preview release for Amazon Connect and is subject to change.

ListInstanceStorageConfigs

This API is in preview release for Amazon Connect and is subject to change.

ListInstances

This API is in preview release for Amazon Connect and is subject to change.

ListIntegrationAssociations

Provides summary information about the AppIntegration associations for the specified Amazon Connect

ListLambdaFunctions

This API is in preview release for Amazon Connect and is subject to change.

ListLexBots

This API is in preview release for Amazon Connect and is subject to change.

ListPhoneNumbers

Provides information about the phone numbers for the specified Amazon Connect instance.

ListPrompts

Provides information about the prompts for the specified Amazon Connect instance.

ListQueueQuickConnects

This API is in preview release for Amazon Connect and is subject to change.

ListQueues

Provides information about the queues for the specified Amazon Connect instance.

ListQuickConnects

Provides information about the quick connects for the specified Amazon Connect instance.

ListRoutingProfileQueues

Lists the queues associated with a routing profile.

ListRoutingProfiles

Provides summary information about the routing profiles for the specified Amazon Connect

ListSecurityKeys

This API is in preview release for Amazon Connect and is subject to change.

ListSecurityProfiles

Provides summary information about the security profiles for the specified Amazon Connect

ListTagsForResource

Lists the tags for the specified resource.

ListUseCases

Lists the use cases.

ListUserHierarchyGroups

Provides summary information about the hierarchy groups for the specified Amazon Connect

ListUsers

Provides summary information about the users for the specified Amazon Connect instance.

ResumeContactRecording

When a contact is being recorded, and the recording has been suspended using

StartChatContact

Initiates a contact flow to start a new chat for the customer. Response of this API provides

StartContactRecording

Starts recording the contact when the agent joins the call. StartContactRecording is a

StartOutboundVoiceContact

Places an outbound call to a contact, and then initiates the contact flow. It performs the

StartTaskContact

Initiates a contact flow to start a new task.

StopContact

Ends the specified contact.

StopContactRecording

Stops recording a call when a contact is being recorded. StopContactRecording is a one-time

SuspendContactRecording

When a contact is being recorded, this API suspends recording the call. For example, you

TagResource

Adds the specified tags to the specified resource.

UntagResource

Removes the specified tags from the specified resource.

UpdateAgentStatus

This API is in preview release for Amazon Connect and is subject to change.

UpdateContactAttributes

Creates or updates

UpdateContactFlowContent

Updates the specified contact flow.

UpdateContactFlowName

The name of the contact flow.

UpdateHoursOfOperation

This API is in preview release for Amazon Connect and is subject to change.

UpdateInstanceAttribute

This API is in preview release for Amazon Connect and is subject to change.

UpdateInstanceStorageConfig

This API is in preview release for Amazon Connect and is subject to change.

UpdateQueueHoursOfOperation

This API is in preview release for Amazon Connect and is subject to change.

UpdateQueueMaxContacts

This API is in preview release for Amazon Connect and is subject to change.

UpdateQueueName

This API is in preview release for Amazon Connect and is subject to change.

UpdateQueueOutboundCallerConfig

This API is in preview release for Amazon Connect and is subject to change.

UpdateQueueStatus

This API is in preview release for Amazon Connect and is subject to change.

UpdateQuickConnectConfig

Updates the configuration settings for the specified quick connect.

UpdateQuickConnectName

Updates the name and description of a quick connect. The request accepts the following data in JSON format. At least Name or Description must be provided.

UpdateRoutingProfileConcurrency

Updates the channels that agents can handle in the Contact Control Panel (CCP) for a routing

UpdateRoutingProfileDefaultOutboundQueue

Updates the default outbound queue of a routing profile.

UpdateRoutingProfileName

Updates the name and description of a routing profile. The request accepts the following data in JSON format.

UpdateRoutingProfileQueues

Updates the properties associated with a set of queues for a routing profile.

UpdateUserHierarchy

Assigns the specified hierarchy group to the specified user.

UpdateUserHierarchyGroupName

Updates the name of the user hierarchy group.

UpdateUserHierarchyStructure

Updates the user hierarchy structure: add, remove, and rename user hierarchy levels.

UpdateUserIdentityInfo

Updates the identity information for the specified user.

UpdateUserPhoneConfig

Updates the phone configuration settings for the specified user.

UpdateUserRoutingProfile

Assigns the specified routing profile to the specified user.

UpdateUserSecurityProfiles

Assigns the specified security profiles to the specified user.